net,ovs: Remove OVS CNI tests
OVS CNI is not used anymore, and need not be covered.
This removal will also contribute to the tests stability, having less flaky tests.
Note that this step doesn't mean OVS bridge is not used and tested anymore; it is still
used for localnet bridging, but for that a different CNI is used.
