+++ tests/conftest.py
@@ -149,7 +149,6 @@ from utilities.infra import (
     get_infrastructure,
     get_node_selector_dict,
     get_nodes_with_label,
-    get_pods,
     get_subscription,
     get_utility_pods_from_nodes,
     label_nodes,
@@ -157,19 +156,15 @@ from utilities.infra import (
     login_with_user_password,
     run_virtctl_command,
     scale_deployment_replicas,
-    wait_for_pods_deletion,
 )
 from utilities.network import (
     EthernetNetworkConfigurationPolicy,
     MacPool,
     cloud_init_network_data,
-    enable_hyperconverged_ovs_annotations,
     get_cluster_cni_type,
     network_device,
     network_nad,
     wait_for_node_marked_by_bridge,
-    wait_for_ovs_daemonset_resource,
-    wait_for_ovs_status,
 )
 from utilities.operator import (
     cluster_with_icsp,
@@ -195,7 +190,6 @@ from utilities.virt import (
     fedora_vm_body,
     get_base_templates_list,
     get_hyperconverged_kubevirt,
-    get_hyperconverged_ovs_annotations,
     get_kubevirt_hyperconverged_spec,
     kubernetes_taint_exists,
     running_vm,
@@ -1223,16 +1217,6 @@ def kubevirt_feature_gates_scope_module(kubevirt_config_scope_module):
     return kubevirt_config_scope_module["developerConfiguration"][FEATURE_GATES]
 
 
-@pytest.fixture(scope="class")
-def ovs_daemonset(admin_client, hco_namespace):
-    return wait_for_ovs_daemonset_resource(admin_client=admin_client, hco_namespace=hco_namespace)
-
-
-@pytest.fixture()
-def hyperconverged_ovs_annotations_fetched(hyperconverged_resource_scope_function):
-    return get_hyperconverged_ovs_annotations(hyperconverged=hyperconverged_resource_scope_function)
-
-
 @pytest.fixture(scope="session")
 def network_addons_config_scope_session(admin_client):
     nac = list(NetworkAddonsConfig.get(client=admin_client))
@@ -1257,31 +1241,6 @@ def skip_test_if_no_ocs_sc(ocs_storage_class):
     """
     if not ocs_storage_class:
         pytest.skip("Skipping test, OCS storage class is not deployed")
-
-
-@pytest.fixture(scope="session")
-def hyperconverged_ovs_annotations_enabled_scope_session(
-    admin_client,
-    hco_namespace,
-    hyperconverged_resource_scope_session,
-    network_addons_config_scope_session,
-):
-    yield from enable_hyperconverged_ovs_annotations(
-        admin_client=admin_client,
-        hco_namespace=hco_namespace,
-        hyperconverged_resource=hyperconverged_resource_scope_session,
-        network_addons_config=network_addons_config_scope_session,
-    )
-
-    # Make sure all ovs pods are deleted:
-    wait_for_ovs_status(network_addons_config=network_addons_config_scope_session, status=False)
-    wait_for_pods_deletion(
-        pods=get_pods(
-            client=admin_client,
-            namespace=hco_namespace,
-            label="app=ovs-cni",
-        )
-    )
 
 
 @pytest.fixture(scope="session")
