+++ tests/network/bond/test_bond_modes.py
@@ -8,7 +8,7 @@ from contextlib import contextmanager
 
 import pytest
 
-from utilities.constants import TIMEOUT_9MIN
+from utilities.constants import LINUX_BRIDGE, TIMEOUT_9MIN
 from utilities.infra import ExecCommandOnPod, get_node_selector_dict, get_node_selector_name
 from utilities.network import (
     BondNodeNetworkConfigurationPolicy,
@@ -20,7 +20,6 @@ from utilities.virt import VirtualMachineForTests, fedora_vm_body
 pytestmark = [
     pytest.mark.sno,
     pytest.mark.usefixtures(
-        "hyperconverged_ovs_annotations_enabled_scope_session",
         "workers_type",
     ),
 ]
@@ -86,10 +85,10 @@ def matrix_bond_modes_bond(
 
 
 @pytest.fixture()
-def bond_modes_nad(admin_client, bridge_device_matrix__function__, namespace, matrix_bond_modes_bond):
+def bond_modes_nad(admin_client, namespace, matrix_bond_modes_bond):
     with network_nad(
         namespace=namespace,
-        nad_type=bridge_device_matrix__function__,
+        nad_type=LINUX_BRIDGE,
         nad_name=f"bond-nad-{matrix_bond_modes_bond.bond_name}",
         interface_name=f"br{matrix_bond_modes_bond.bond_name}",
         client=admin_client,
@@ -101,7 +100,6 @@ def bond_modes_nad(admin_client, bridge_device_matrix__function__, namespace, ma
 def matrix_bond_modes_bridge(
     nmstate_dependent_placeholder,
     admin_client,
-    bridge_device_matrix__function__,
     worker_node1,
     bond_modes_nad,
     matrix_bond_modes_bond,
@@ -110,7 +108,7 @@ def matrix_bond_modes_bridge(
     Create bridge and attach the BOND to it
     """
     with network_device(
-        interface_type=bridge_device_matrix__function__,
+        interface_type=LINUX_BRIDGE,
         nncp_name=f"bridge-on-bond-{matrix_bond_modes_bond.bond_name}",
         node_selector=get_node_selector_dict(node_selector=worker_node1.hostname),
         interface_name=bond_modes_nad.bridge_name,
@@ -141,7 +139,6 @@ def bond_modes_vm(
 def bridge_on_bond_fail_over_mac(
     nmstate_dependent_placeholder,
     admin_client,
-    bridge_device_matrix__function__,
     worker_node1,
     bond_modes_nad,
     active_backup_bond_with_fail_over_mac,
@@ -150,7 +147,7 @@ def bridge_on_bond_fail_over_mac(
     Create bridge and attach the BOND to it
     """
     with network_device(
-        interface_type=bridge_device_matrix__function__,
+        interface_type=LINUX_BRIDGE,
         nncp_name="bridge-on-bond-fail-over-mac",
         node_selector=get_node_selector_dict(node_selector=worker_node1.hostname),
         interface_name=bond_modes_nad.bridge_name,
