+++ tests/network/jumbo_frame/test_bridge.py
@@ -9,6 +9,7 @@ import pytest
 from libs.net.ip import random_ipv4_address
 from libs.net.vmspec import lookup_iface_status_ip
 from tests.network.utils import assert_no_ping
+from utilities.constants import LINUX_BRIDGE
 from utilities.infra import get_node_selector_dict
 from utilities.network import (
     assert_ping_successful,
@@ -19,9 +20,6 @@ from utilities.network import (
 from utilities.virt import VirtualMachineForTests, fedora_vm_body
 
 pytestmark = [
-    pytest.mark.usefixtures(
-        "hyperconverged_ovs_annotations_enabled_scope_session",
-    ),
     pytest.mark.special_infra,
     pytest.mark.jumbo_frame,
 ]
@@ -37,13 +35,12 @@ def jumbo_frame_bridge_device_worker_1(
     nmstate_dependent_placeholder,
     admin_client,
     cluster_hardware_mtu,
-    bridge_device_matrix__class__,
     worker_node1,
     hosts_common_available_ports,
     jumbo_frame_bridge_device_name,
 ):
     with network_device(
-        interface_type=bridge_device_matrix__class__,
+        interface_type=LINUX_BRIDGE,
         nncp_name="jumbo-frame-bridge-nncp-1",
         interface_name=jumbo_frame_bridge_device_name,
         node_selector=get_node_selector_dict(node_selector=worker_node1.hostname),
@@ -59,13 +56,12 @@ def jumbo_frame_bridge_device_worker_2(
     nmstate_dependent_placeholder,
     admin_client,
     cluster_hardware_mtu,
-    bridge_device_matrix__class__,
     worker_node2,
     hosts_common_available_ports,
     jumbo_frame_bridge_device_name,
 ):
     with network_device(
-        interface_type=bridge_device_matrix__class__,
+        interface_type=LINUX_BRIDGE,
         nncp_name="jumbo-frame-bridge-nncp-2",
         interface_name=jumbo_frame_bridge_device_name,
         node_selector=get_node_selector_dict(node_selector=worker_node2.hostname),
@@ -80,7 +76,6 @@ def jumbo_frame_bridge_device_worker_2(
 def br1test_bridge_nad(
     admin_client,
     cluster_hardware_mtu,
-    bridge_device_matrix__class__,
     namespace,
     jumbo_frame_bridge_device_name,
     jumbo_frame_bridge_device_worker_1,
@@ -88,7 +83,7 @@ def br1test_bridge_nad(
 ):
     with network_nad(
         namespace=namespace,
-        nad_type=bridge_device_matrix__class__,
+        nad_type=LINUX_BRIDGE,
         nad_name=f"{jumbo_frame_bridge_device_name}-nad",
         interface_name=jumbo_frame_bridge_device_name,
         mtu=cluster_hardware_mtu,
