+++ tests/network/l2_bridge/test_l2_linux_bridge.py
@@ -8,22 +8,20 @@ from timeout_sampler import TimeoutSampler
 from libs.net.vmspec import lookup_iface_status_ip
 from tests.network.l2_bridge.libl2bridge import wait_for_no_packet_loss_after_connection
 from tests.network.libs.dhcpd import DHCP_IP_RANGE_START
-from utilities.constants import LINUX_BRIDGE, OVS_BRIDGE, TIMEOUT_2MIN
+from utilities.constants import TIMEOUT_2MIN
 from utilities.network import assert_ping_successful
 
-pytestmark = [pytest.mark.ipv4, pytest.mark.usefixtures("hyperconverged_ovs_annotations_enabled_scope_session")]
+pytestmark = [pytest.mark.ipv4]
 
 CUSTOM_ETH_PROTOCOL = "0x88B6"  # rfc5342 Local Experimental Ethertype. Used to test custom eth type
 
 
-@pytest.mark.parametrize("bridge_device_type", [LINUX_BRIDGE, OVS_BRIDGE], indirect=True)
 @pytest.mark.s390x
 class TestL2Bridge:
     """
-    Test L2 connectivity via Linux or OVS bridge CNI plugin.
-    Each bridge is configured by fixtures with matrix.
+    Test L2 connectivity via Linux bridge CNI plugin.
     The main goal is to make sure that different kinds of L2 traffic can pass
-    transparently via Linux/OVS Bridge.
+    transparently via Linux Bridge.
     """
 
     @pytest.mark.polarion("CNV-2285")
